requirements.txt
Pillow
escpos
spacy
markovify
requests
